- 지하철 노선 관련 기능을 다른 클래스로 분리
